v 0.2.3
- Service Principal Fix
- Project Reorganization
